deployCollectData/spamCollect/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
phonenumbers==8.13.20
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
